#27939 Adding storybook for Prime ng table component
