+++ core-web/libs/dotcms-scss/angular/dotcms-theme/theme.scss
@@ -33,6 +33,7 @@
 @use "components/toolbar";
 @use "components/tooltip";
 @use "components/tree";
+@use "components/table";
 
 @use "utils/validation";
 @use "utils/password";
