--- core-web/libs/ui/src/lib/dot-table/dot-table.component.html
@@ -1,52 +1,0 @@
-<p-table [value]="products" styleClass="dot-table" dataKey="code">
-    <ng-template pTemplate="header">
-        <tr>
-            <th style="width: 4rem">
-                <p-tableHeaderCheckbox></p-tableHeaderCheckbox>
-            </th>
-            <th pSortableColumn="name">Title <p-sortIcon field="name"></p-sortIcon></th>
-            <th pSortableColumn="status">Status <p-sortIcon field="status"></p-sortIcon></th>
-            <th pSortableColumn="assignee">Assignee <p-sortIcon field="assignee"></p-sortIcon></th>
-            <th pSortableColumn="date">Last Updated <p-sortIcon field="date"></p-sortIcon></th>
-            <th pSortableColumn="date">Menu</th>
-        </tr>
-    </ng-template>
-    <ng-template pTemplate="body" let-product>
-        <tr>
-            <td>
-                <p-tableCheckbox [value]="product"></p-tableCheckbox>
-            </td>
-            <td>{{ product.name }}</td>
-            <td>
-                <ng-container [ngSwitch]="product.status.toLowerCase()">
-                    <p-tag
-                        class="sm p-tag-success"
-                        *ngSwitchCase="'published'"
-                        [ngClass]="status.published">
-                        {{ product.status }}
-                    </p-tag>
-                    <p-tag
-                        class="sm p-tag-warning"
-                        *ngSwitchCase="'draft'"
-                        [ngClass]="status.archived">
-                        {{ product.status }}
-                    </p-tag>
-                    <p-tag class="sm p-tag-info" *ngSwitchCase="'new'" [ngClass]="status.new">
-                        {{ product.status }}
-                    </p-tag>
-                    <p-tag class="sm p-tag-error" *ngSwitchCase="'archived'" [ngClass]="status.new">
-                        {{ product.status }}
-                    </p-tag>
-                    <p-tag class="sm p-tag" *ngSwitchDefault>
-                        {{ product.status }}
-                    </p-tag>
-                </ng-container>
-            </td>
-            <td>{{ product.assignee }}</td>
-            <td>{{ product.date | date }}</td>
-            <td>
-                <i class="pi pi-ellipsis-v" style="color: #426bf0"></i>
-            </td>
-        </tr>
-    </ng-template>
-</p-table>
