--- core-web/libs/ui/src/lib/dot-table/dot-table.component.scss
@@ -1,31 +1,0 @@
-@use "variables" as *;
-
-:host {
-    border: 3px solid $color-palette-gray-400;
-    width: 100%;
-    display: flex;
-    border-radius: 5px;
-}
-
-::ng-deep {
-    p-table {
-        width: 100%;
-    }
-
-    .p-datatable .p-datatable-thead > tr > th {
-        background-color: $color-palette-gray-200;
-        color: black;
-    }
-
-    .dot-table .p-checkbox {
-        border: 3px solid $color-palette-gray-400;
-    }
-
-    .p-datatable .p-sortable-column:focus {
-        background-color: $color-palette-gray-200;
-    }
-
-    .p-datatable .p-sortable-column.p-highlight {
-        background-color: $color-palette-gray-200;
-    }
-}
