--- core-web/libs/ui/src/lib/dot-table/dot-table.component.stories.ts
@@ -1,21 +1,0 @@
-import { moduleMetadata } from '@storybook/angular';
-
-import { CommonModule } from '@angular/common';
-
-import { TableModule } from 'primeng/table';
-
-import { DotTableComponent } from './dot-table.component';
-
-export default {
-    title: 'DotTableComponent',
-    component: DotTableComponent,
-    decorators: [
-        moduleMetadata({
-            imports: [TableModule, CommonModule]
-        })
-    ]
-};
-
-export const Default = () => ({
-    component: DotTableComponent
-});
